flask_app/requirements.txt
Flask
Flask-WTF
Flask-SQLAlchemy
psycopg2-binary